Improve Web API Course
# Change 1
The article covering testing with Swagger didn't mention that the launchsettings.json file has to be configured with the appropriate properties in order for the app to open a browser with the swagger url when it runs. This had me confused when I went over the article and read that my app should open it but it didn't. A paragraph along with a code snippet has been added to the "Testing with Swagger" section before it asks to run the app.

# Change 2
Fixed typos, grammar, and punctuation in the Conclusion article's text content