feat: add method to retrieve recently joined users
Updated the `IUserService` interface to include a new method, `GetRecentlyJoinedUsers(int count)`, which returns a list of recently joined `ApplicationUser` objects.

Implemented this method in the `UserService` class to query the `AspNetUsers` table, ordering by `CreatedDate` and limiting results to the specified count. Added error handling to log exceptions and return an empty list in case of errors.