refactor: move nav links to not authorized section
The code changes in `NavMenu.razor` involve the removal of several `<MudNavLink>` components that link to "Community Feed," "Articles," and "Blog" for authorized users. These links have been added to the not authorized section, allowing unauthenticated users to access them. This change enhances navigation accessibility for users who are not logged in.